# services/api-gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Unified Task Manager API
  version: 0.1.0
  description: |
    Public HTTP API exposed through the API Gateway (`/api` prefix on the edge).
    Nginx strips `/api` before proxying; paths below are relative to that prefix.

    | Prefix | Service |
    |--------|---------|
    | `/v1/auth`, `/v1/users`, `/v1/teams`, `/v1/permissions` | user-service |
    | `/v1/tasks`, `/v1/task-columns`, `/v1/boards` | task-service |
    | `/v1/chats` | chat-service |
    | `/v1/projects`, `/v1/tasks/.../deadline-notification` | notification-service |
    | `/v1/file-environments` | file-service |
    | `/api/tasks/assistant`, `/task_name_description` | ml-service |
    | `/calls` | sfu-service (WebRTC signaling at `/ws/calls` on the edge) |

    **Auth:** `Authorization: Bearer <accessToken>` except public auth routes.
    **Team scope:** send `X-Team-Id` for task/column APIs (gateway also accepts `teamId` query on board WebSocket).
    **Optimistic concurrency:** task/column PATCH accepts `If-Match` (ETag from GET/PATCH); `412` returns `VERSION_CONFLICT` and `current` entity.
servers:
  - url: /api
    description: API Gateway (browser clients use this base URL)
...
security:
  - bearerAuth: []
paths:
...
  /v1/file-environments/ensure:
    post:
      tags: [Files]
      summary: Ensure file environment exists for project
      operationId: ensureFileEnvironment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileEnvironmentEnsureRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileEnvironment'
  /v1/file-environments/members:
    post:
      tags: [Files]
      summary: Add members to file environment
      operationId: addFileEnvironmentMembers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileMembersRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      type: string
    delete:
      tags: [Files]
      summary: Remove members from file environment
      operationId: removeFileEnvironmentMembers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileMembersRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      type: string
  /v1/file-environments/folders:
    post:
      tags: [Files]
      summary: Create folder
      operationId: createFileFolder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilePathRequest'
  /v1/file-environments/files:
    post:
      tags: [Files]
      summary: Upload file
      operationId: uploadFile
      parameters:
        - name: team_id
          in: query
          required: true
          schema:
            type: string
        - name: project_id
          in: query
          required: true
          schema:
            type: string
        - name: directory
          in: query
          schema:
            type: string
            default: .
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
  /v1/file-environments/entries:
    get:
      tags: [Files]
      summary: List directory entries
      operationId: listFileEntries
      parameters:
        - name: team_id
          in: query
          required: true
          schema:
            type: string
        - name: project_id
          in: query
          required: true
          schema:
            type: string
        - name: directory
          in: query
          schema:
            type: string
            default: .
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileEntryListResponse'
    delete:
      tags: [Files]
      summary: Delete file or folder
      operationId: deleteFileEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilePathRequest'
  /v1/file-environments/entries/rename:
    patch:
      tags: [Files]
      summary: Rename file or folder
      operationId: renameFileEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileRenameRequest'
  /v1/file-environments/files/download:
    get:
      tags: [Files]
      summary: Download file
      operationId: downloadFile
      parameters:
        - name: team_id
          in: query
          required: true
          schema:
            type: string
        - name: project_id
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /v1/file-environments/files/view:
    get:
      tags: [Files]
      summary: View file inline
      operationId: viewFile
      parameters:
        - name: team_id
          in: query
          required: true
          schema:
            type: string
        - name: project_id
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bearerQuery:
      type: apiKey
      in: query
      name: token
      description: Bearer access token in query (WebSocket clients)
...
  schemas:
...
    FileEnvironmentEnsureRequest:
      type: object
      required: [team_id, project_id]
      properties:
        team_id: { type: string }
        project_id: { type: string }
    FileEnvironment:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        owner_user_id: { type: string }
        team_id: { type: string }
        project_id: { type: string }
        member_user_ids:
          type: array
          items: { type: string }
    FileMembersRequest:
      type: object
      required: [team_id, project_id, user_ids]
      properties:
        team_id: { type: string }
        project_id: { type: string }
        user_ids:
          type: array
          items: { type: string }
          minItems: 1
    FilePathRequest:
      type: object
      required: [team_id, project_id, path]
      properties:
        team_id: { type: string }
        project_id: { type: string }
        path: { type: string }
    FileRenameRequest:
      type: object
      required: [team_id, project_id, old_path, new_path]
      properties:
        team_id: { type: string }
        project_id: { type: string }
        old_path: { type: string }
        new_path: { type: string }
    FileEntry:
      type: object
      properties:
        name: { type: string }
        path: { type: string }
        kind: { type: string, enum: [file, directory] }
        size_bytes: { type: integer }
    FileEntryListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FileEntry'
...